requirements.txt
streamlit==1.50.0
pandas==2.3.3
gspread==6.1.4
google-auth==2.35.0